refactor(schema) - standardize schemas folder
